Enable Datadog APM
In order to enable Datadog APM in Heroku the Docker image needs to open
ports and install the agent binary.

To start the agent binary, a docker entrypoint is created and will run
in the background the agent's process before executing the command send
to the container.

More info: https://docs.datadoghq.com/agent/basic_agent_usage/heroku/