Don't make cron sentence mandatory
The cron sentence should not be mandatory, allowing executions to be
only manually.

This will avoid forcing users to specify a cron sentence if they don't
want to and will reduce pressure on the execution scheduler by reducing
by default the number of jobs to run.

Lastly, this now opens the door to create an endpoint to trigger
executions via webhooks.