Match non valid execution status and halt them
Currently there is a monolith of this application. This means any worker
executing a job when a new version is deployed means it will die and let
the execution in a *inconsistent running state*.

After a deploy, the started workers will try to restart the job again,
but that can lead to unexpected side effects to users.

To cover the case mentioned above, every time a worker tries to retake a
execution job and checks the current status, if it's invalid, will halt
the execution.