Fix infinite limit expiration time
If a limit cache is not set, set the expiration time. Otherwise keep the
current one and wait until is gone to set a new one.

This provides a "throttle" where during the cache time since the first
execution all the other ones are collected until the limit is done.